oled的i2c地址通过设备配置
原来直接写在bus_lcd文件中的函数

+++ Petite/bus/bus_i2c.h
@@ -34,7 +34,7 @@ typedef struct
 }DevI2cNode;
 
 typedef struct{
-	uint16_t addr;
+	uint8_t addr;
 	uint16_t fkhz;
 }I2CPra;
